flask_app/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy
gunicorn==20.1.0
marshmallow==3.14.0
psycopg2-binary
#psycopg2==2.8.5
PyJWT==2.3.0
redis==3.3.8
SQLAlchemy==1.3.24
kombu==4.6.11
amqp==2.6.1
celery==4.4.6